Split video source again
Allow the user to control whether or not to start the local video
capture by controlling a `LocalVideoSource` component.
Allow the user to optionally render that local video feed via a
video track player referencing that local video source.
Allow the user to control whether or not to add listeners to signals
indicating a new remote video stream was added/removed via
a `RemoteVideoSource` component.
Allow the user to optionally render that remote video feed via a
video track player referencing that remote video source.

This enables scenarios like:
- Send local video without render
- Receive remote video into custom sink (no display, or custom
  render pipeline not using the `VideoTrackPlayer` component)
- Send multiple local video tracks by adding multiple local video
  source components referencing different video capture devices.
- Received and render (or custom process) multiple remote video
  tracks by adding multiple remote video source components
  referencing different tracks.

TODO:
- Add support for enumerating and selecting a local video source
  from `LocalVideoSource`, including inside the Unity editor.
- Add support for referencing a remote track before it is received,
  possibly by ID?