Sync webrtc-uwp-sdk to 134c00e for latest changes
Bump the webrtc-uwp-sdk dependency reference to latest commit in
releases/m71 to integrate recent changes, in particular the --cpp17 flag
change required to compile the project.

This fixes an issue where following the instructions in the README.md to
checkout the project would yield an earlier version of the dependency
without that change, resulting in a build error.