Tentative fix for TestAppUWP missing DLL
Set `CopyToOuptutDirectory=Never` for the native DLL as referenced in
the `TestAppUWP` project. The native DLL is already output to the
correct location by its project, and the `CopyToOutputDirectory` does
not seem to have any influence on AppX deployment (the native DLL is
still copied to the deployment folder). So avoid asking MSBuild any
extra copy, which might be why despite of the project build order it
fails to find the native DLL when starting to compile the `TestAppUWP`
project.