Create the media source with the correct format
Ensure that the local media source is created based on the actual
resolution of the WebRTC frame.

For the moment this is approximated as the resolution of the first
format of the selected video profile, and there is no fallback if no
profile is selected. This is because there is currently no mechanism to
retrieve the format of the WebRTC feed after the video capture device is
open.

Note that this currently crash in Visual Studio the second time the
local video feed start, when setting localVideoPlayer.Source = null.