Refactor DataChannel and interop
Add a new `DataChannel` C++ object, and refactor interop to match it
with its C# counterpart directly. This cleans up the C++ API which now
returns a `DataChannel` object instead of an opaque handle.

Start to move interop-related code into a separate set of files, to
clean up `PeerConnection.cs` and make it more readable, while allowing
possibly later a split into per-platform libraries if this is found to
be the only solution to support .NET Core 3.0 properly.