Add missing data channel callbacks
Add missing callbacks when data channel is created from a remote peer
and negotiated in-band.
Add C# events `DataChannelAdded` and `DataChannelRemoved`, reliably
invoked when any data channel is added or removed, whether in-band or
out-of-band created, and whether local or remote.
Update TestAppUWP to support multiple text chats in parallel as an
example of use, with the default chat being out-of-band and the other
ones added later being in-band.