Thread-safe peer connection factory management
Ensure that operations on the global peer connection factory are
thread-safe, to allow multiple peer connections to to be created in
parallel with the same factory, which will attempt to lazily initialize
it from multiple threads. This allows e.g. using multiple
`PeerConnection` components in Unity and initialize them all at startup.

Bug: #81