Add .NET Core console app sample
Add a sample signaler `NamedPipeSignaler` based as the name implies on
named pipes, and written in .NET Standard 2.0. The signaler works out of
the box without configuration for local connection between two peers on
the same machine.

Add a sample .NET Core 3.0 console app based on that signaler to demo
the use of the C# library.