Keep RTP senders alive for local tracks
Previously local audio and video tracks were creating a new RTP sender
and transceiver each time they were added to the local peer connection.
Because RTP transceivers cannot be removed, this produces piles of dead
objects that cannot be reclaimed until the peer connection is destroyed.

This change keeps the RTP senders alive, and reuse them and their
underlying RTP transceiver, by dynamically swapping the sender's track
to NULL on remove and restoring it to the actual local track on re-add.
This allows reusing both the RTP sender and the RTP transceiver when
adding/removing the local audio and video tracks multiple times.