Move interop API to private, expose public C++ API
Remove the `api.h` file from the public API and rename to
`interop_api.h` to avoid any confusion. This is an internal API for
interop with C#, that should not be used directly.

Expose an actual C++ API by moving the headers of the C++ classes to the
`include/` folder.