Expose PeerConnectionInterface::SetBitrate (#109)
Expose webrtc::PeerConnectionInterface::SetBitrate() to allow the user to set
the minimum and maximum bitrates used for the RTP tracks, as well as the
starting value. This enables in particular changing the default video value of
300 kpbs to something higher in local scenarios where the bandwidth is expected
to be high, to enable higher quality video right from the start of the video
track recording.