Remove '.' from Unity project name
This works around a deploy bug in Unity where the presence of '.' in the
project name sometimes confuses the Build script to generate invalid
projects. Symptoms are a wrong _Data folder location and failure to run.
This is more often observed in Desktop PC builds. Bug reported to Unity
which is fixing.