Add C# tracing utility for ETW event recording
Add a new `MainEventSource` C# class holding a static singleton instance
`Log` used to report events which can be recorded as ETW events. This
enables developers to record traces for performance analysis and for
debugging.

An ETW recording profile (*.wprp) is provided in `tools/tracing/` for
users to record their own trace.