Make PeerConnection closeable without destroying it
Ensure that PeerConnection::Close() is closing the peer connection and
releasing the internal implementation object, without desotrying the
actual `PeerConnection` C++ instance. This simplifies the clean-up
process, which requires that internal WebRTC implementation objects be
destroyed on the WebRTC signaling thread, something that is difficult to
guarantee if destroyed at the same time as the `PeerConnection` C++
object instance itself.

This is also a preparatory work for upcoming API changes introducing
more standalone objects which can outlive the peer connection, by
introducing a GlobalFactory object keeping track of all the objects
alive to decide when to release the global WebRTC factory and threads.