Fix stats calculation in VideoFrameQueue
- Ensure that the average stat per second is calculated as the inverse
of the delta time average, and not the average of the inverse delta
time, to produce the arithmetic mean instead of the harmonic mean.
- Push some values every frame to the dropped queue, even if no frame
was dropped, to ensure the stat is updated at least once per read.
Otherwise with very low frequency push (nearly no dropped frame) the
stat queue is seldomly updated and very old values participate in the
average, which does not represent anymore an average over the last
second or so.