Fixes for #132 (cherry-pick from release/1.0)
Cherry-pick an issue fixed while back-porting #132 from master to
release/1.0.

The H.264 encoder/decoder code is only available on UWP, but is compiled
on Desktop platforms, although it requires some UWP-specific libraries
to be linked against. Under normal linker stripping conditions this is
not an issue, but with the previous change the extern global was pulling
a UWP-specific module, making the link fail. This change also fix that issue
compared to the cheery-pick.