Backport #132 HL1 frame height workaround (#137)
Cherry-pick changes of #132

Backport workaround for the HoloLens 1 H.264 encoder bug producing
artifacts when the frame height is not a multiple of 16 pixels. These
changes allow the app to select a padding or cropping behavior to work
around the bug. On HoloLens 1, the behavior defaults to cropping.
Otherwise it defaults to no-op.

The H.264 encoder/decoder code is only available on UWP, but is compiled
on Desktop platforms, although it requires some UWP-specific libraries
to be linked against. Under normal linker stripping conditions this is
not an issue, but with the previous change the extern global was pulling
a UWP-specific module, making the link fail. This change also fix that issue
compared to the cheery-pick.