Keep the correct Delegate instance alive
Ensure that delegates registered as callbacks to the native layer are
kept alive, preventing the Garbage Collector from destroying them while
still registered as callbacks.

Previously a first Delegate instance was implicitly created when
assigning to the callback arguments, and kept alive, but then the method
identifier was used to call the P/Invoke registration, which implicitly
created a second Delegate instance and registered that one, without
keeping any reference to it. On GC, this would crash with a
NullReferenceException (best case) or ExecutionEngineException (worst
case), making this very difficult to diagnose.

Note that this seems to solve 2 known intermitent crashes:
- Crash when enumerating video devices and video device formats;
- Crash when using external video tracks from C#.