Split external source creation from track add
Split the process of creating an external video track source from the
process of adding a new local video track based on that source. This
makes things more clear.

Ensure that the external video track C# wrapper is created when the
native frame callback is registered, as the native external video track
source immediately starts to capture, and the wrapper is needed to
complete the frame request.

Add some C# tests for I420A and ARGB32 callbacks based external video
track sources.