Invert Unity shader UV vertical convention
Previously the Unity shaders for YUV and ARGB feeds were assuming a
DirectX model where the V coordinate starts at the top of the texture,
and reversing it if UNITY_UV_STARTS_AT_TOP was false. Although this
works well with the built-in Unity plane mesh, this is unintuitive when
used with a standard quad in cartesian coordinates with X right and Y
up, where UV start at (0,0) in the lower left corner at (-0.5,-0.5), and
go to (1,1) in the upper right corner at (+0.5,+0.5), assuming a unit
quad.

This change reverse the convention so that it intuitively maps to that
representation, and inverse the V coordinate when UNITY_UV_STARTS_AT_TOP
is true. This maps directly to the built-in Unity quad mesh, which
should be preferred to the built-in plane for rendering a simple
rectangle containing a video feed. Custom meshes might need to be
adapted or the change manually reverted if users previously assumed the
same convention as those shaders.