Discard old frame requests in external video source
This change discards old pending frame requests in an external video
track source to fit new ones instead of skipping newer requests. This
allows the frame pipeline to "restart" after a long delay, and not get
stuck in an infinite loop of skipped frames, as most commonly users will
only complete a frame request (and delete the request) once receiving a
new request.