Set stereoscopic rendering to Single-Pass
Default stereoscopic rendering mode to Single-Pass (non-instanced) even
though this is less performant than Single-Pass Instanced, because the
latter does not allow SceneVideoSource to work on Unity 2018.4 due to a
missing CommandBuffer API to capture a single texture array layer of the
camera back-buffer in this mode. Users can still (and are recommended
to) use Single-Pass Instanced if not using SceneVideoSource, or if using
Unity 2019.1+ where the relevant API is available, noting thought that
Unity 2019.x is not officially supported by MixedReality-WebRTC, despite
the fact there is no known issue with it at this time.