Add Unity integration for external video tracks
Add some Unity integration of the external video tracks feature by way
of a CustomVideoSource component allowing the user to inject some custom
video feed in the WebRTC connection.

As an example application, provide a SceneVideoSource component
capturing the content rendered by a given Unity camera and streaming it
as a video feed to the remote peer.

The VideoChatDemo Unity scene is updated to add a scene capture view
showing via a MediaPlayer component the scene content captured from the
camera back-buffer.

Bug: #35