Backport autogenerated list of documented branches
Automatically generate a list of git branches for which a documentation
exists, and export it to a JavaScript file to be read by the branch
selector script. The branch list is saved to `styles/branches.gen.js`
inside the `_docs` folder during documentation building.

Also add the name of the current branch the documentation is for, which
allow updating the selection of the branch <select> item.

Cherry-pick: 113fd6d