Preserve include blocks with clang-format
The clang-format configuration uses the Chromium style, which according
to
https://chromium.googlesource.com/chromium/src/+/master/styleguide/c++/c++.md
seems to correspond to the IncludeBlocks:Regroup option. Until now, for
some reason, Visual Studio did not seem to enforce that, but instead use
some formatting closer to Preserve. This behavior seems to have changed
around 16.4.2 or so.

This change forces IncludeBlocks:Preserve to prevent clang-format from
reordering C++ includes, which otherwise breaks compiling as the
precompiled header is often moved away and not anymore the first
include. This option ensures that it stays the first include, and blocks
of headers are not reordered. Headers are still sorted within a block.