Improve the codec filtering mechanism
- Add support for setting extra parameters on local codecs. Since there seems to
 be no easier way, we filter remote SDP messages before delivery.
- Allow endpoints which receive offers to filter codecs/request codec parameters
 to the counterpart (before this change, only endpoints creating an offer did
 SDP filtering).

Note that we now filter outgoing SDP answers as well as offers. Before this
change, we didn't filter answers, because removing codecs from an answer without
removing them from the corresponding offer left the connection in an
inconsistent state (e.g. A sends an offer to B with VP8 and H264; B internally
chooses VP8, replies with both codecs, but the answer is filtered to remove VP8;
A gets the filtered answer and chooses H264). Now we use the same filter on
incoming offers and outgoing answers so there is no possibility of inconsistency.