Use clang-cl for ARM64 core builds
The underlying implementation from Google (libwebrtc) cannot be compiled
with MSVC for the ARM64 target. Because this library is statically
linked, and to avoid any compiler mismatch issue, also compile the core
DLLs with the same clang-cl used to build libwebrtc.

This change requires any user recompiling those core DLLs to edit the
Microsoft.MixedReality.WebRTC.Native.targets file to specify where to
find their local repository of Google's libwebrtc.

Incidentally, because libwebrtc and Chromium use a custom version of
libc++ checked in the repository and compiled as part of the normal
library build, and because libwebrtc is statically linked, this change
also requires statically linking against the .obj files of libc++
produced as part of the libwebrtc build.

Note that this change selects the ClangCL toolchain on ARM64, which by
default doesn't exist in Visual Studio and cannot be installed. See the
install_vs_toolchains.md file to modify a local install of Visual Studio
2019 and add the ClangCL toolchain for Desktop and UWP ARM64 platforms.