[Unity] Add missing ARGB unlit shader
Add the missing ARGB32 shader with unlit model, which should more
commonly be used for video chat applications than a lit one.

Set the default color of all shaders to black by modifying the default
textures.

Replace the default shader in the ARGBFeedMaterial to use the unlit
variant of the ARGB shader.