Async SetRemoteDescription to avoid race condition (#195)
Make mrsSetRemoteDescriptionAsync() invoke a callback on completion, and
change the signature of C# methods to be async. This allows the caller
to know when the remote description has been applied, which informs that
it is safe to call CreateAnswer(). Otherwise the call setting the remote
description may race against the call creating the answer, preventing
the answer from being created if the remote description was not applied
yet.

Rename all related methods and functions with the "async" suffix for
readability and to force a breaking change and avoid silently ignoring
the async task, which was an error in existing code.