2-step external video track creation for callbacks (#196)
Split creation of external video tracks into 2 steps, the first creating
the track object, and the second finalizing the creation and starting
the track source capture.

In the current state of the API, the frame callbacks are owned by the
peer connection, so there is little difference. However with the
upcoming change adding standalone track objects, this change allows the
user to register a frame callback before capture starts, and therefore
be able to honor any frame request from the first one, while also
keeping the WebRTC restriction that track sources must start in the
capturing state.