Make GlobalFactory thread-safe to initialize (#197)
Previously RemoveObject() was attempting to destroy the GlobalFactory singleton
instance as soon as the number of tracked objects alive reached zero. This was
problematic for two reasons:

- If thread A retrieves a pointer to the singleton before thread B, then thread
  B creates a new wrapper (which calls AddObject()) but later fails to complete
  the initializing of that object and has to destroy it before thread A itself
  added an object, then RemoveObject() (called from the wrapper destructor)
  will find no object alive and will destroy the GlobalFactory singleton
  instance, which leaves thread A with a dangling pointer.
- Destroying the singleton instance was done outside the lock, and access to
  that pointer was therefore prone to concurrent writes.

This change introduces a different approach where the singleton instance is
created on first use via member function static initializing, which is
thread-safe in C++11, and never destroyed until static de-initializing at the
end of the process or when the module (DLL) is unloaded. Instead, reference
counting and number of tracked objects alive are used to initialize and
shutdown the singleton instance, which now can be reused after shutdown.

The addition of reference counting, and returning a proper
RefPtr<GlobalFactory> instead of a unique pointer, make it easier to understand
for users that the RefPtr instance will prevent the library from shutting down
while it is in scope, and therefore that keeping that RefPtr around while
creating a wrapper object prevents the first issue described above since having
no tracked object alive is necessary but not sufficient anymore to trigger a
library shutdown.