Fix C++ tests for non-ref-owning handles
Handles to native objects are not owning a reference anymore when not
passed from a method creating the object. For example, getting a track
from a transceiver returns a "view" handle that is a copy of the one
previously returned when the track was created, and should *not* be
released while the track is in use. This makes the handle usage more
understandable and similar to a pure C API.