Force native UWP build options for consistency
Enforce the same build flags on the UWP native library than the ones
already existing on the Desktop (Win32) one, for consistency. Note that
in practice this does not change the actual flags used, as those were
already set by default, but ensure consistency and prevent unintended
changes if default values were to change in future Visual Studio
versions.