Remove list of local tracks on peer connection
Local tracks are owned by the user, and do not need to be tracked by the
peer connection. The transceiver they are optionally attached to keeps
them alive while attached, but otherwise there should be no other strong
reference to them. The local track itself keeps a weak back-reference to
the transceiver while attached to it.