Link against core PDBs when building interop DLL (#208)
This change enables linking the interop DLL
Microsoft.MixedReality.WebRTC.Native.dll with the PDBs from the core
webrtc.lib and, on UWP platform, the UWP wrappers from
Org.WebRtc.WrapperGlue.lib, to enable a better debugging experience for
users.

Given the large size of PDBs for each build variant (~400 MB and more),
this changes uses some Universal Packages to store the PDBs in some
private Artifacts feed on Azure DevOps when compiling webrtc.lib, and
restore them when compiling Microsoft.MixedReality.WebRTC.Native.dll, as
the frequency of the former is much lower than the frequency of the
latter, and the former build is generally shared by multiple branches of
the latter (e.g. master and release/1.0 at the moment).

Additionally, because Azure build agents have a limit of 10GB per build,
this change generates on-the-fly a custom packages.config for the
current build triple (config-platform-arch) which allows restoring only
the necessary core NuGet packages (containing webrtc.lib). This has the
extra advantage of reducing the time needed to restore all platform
packages: previously 4 on Desktop and 13 on UWP, now only 1 on Desktop
and 3 on UWP. The .vcxproj is also modified on-the-fly and in-place to
import only the restored packages. Note that those steps are executed
via PowerShell scripts during the build and not stored in the committed
.vcxproj/packages.config files because it is uncertain whether the
Visual Studio IDE would understand platform- and config-dependent NuGet
restore, which are rather unusual and only required because we are
hitting the NuGet file size limit on nuget.org, therefore the files
shipped on GitHub are left untouched with the full set of NuGet packages
for all config/arch per platform, to avoid missing any NuGet restore
which would prevent the user from compiling the solution.

(cherry picked from commit f773f405102f7e90f049f371e2fe3964e3884b1e)