Delayed RenegotiationNeeded to avoid reentrency
Callbacks from native code (reverse P/Invoke) are invoked on the
signaling thread and should not call into the native WebRTC API again.
For the RenegotiationNeeded event, since it is common for the user to
call CreateOffer() as a result of this event, move its call to a worker
thread. In addition, delay the call until the C# Transceiver wrapper has
been fully constructed (and its handle set) to avoid invoking an event
with a partially-constructed object.