Init media_kind with invalid value -1
Initialize mrsTransceiverInitConfig::media_kind with the invalid value
(mrsMediaKind)-1 to catch implementation errors when creating a new
transceiver. Do not introduce an explicit mrsMediaKind::kUnset to avoid
having to handle it everywhere else, as the transceiver constructor will
have validated that argument once already. Note that media_kind is
filled only when directly using the interop DLL; otherwise this is
internal and hidden from the user, so the risk of mistake is limited.