Fix transceiver stream IDs propagation in Plan B
In Unified Plan, transceivers have a guaranteed media line index, as
well as an unlimited number of stream IDs.

In Plan B, there is no transceiver, and no guarantee on the order of
tracks. Moreover tracks have a single stream ID associated with them. To
emulate the features of Unified Plan, this change encodes the media line
index and the list of stream IDs of the emulated transceiver into the
single stream ID of the Plan B track, and decodes them on the remote
peer.