Add support for multi-tracks to Unity integration
Update the Unity integration to changes in the C# API for supporting
multiple tracks per peer connection through the introduction of the
transceiver API.

With this new API, the Unity integration offers an automated way to pair
the transceivers on the local and remote peers based on their index in a
global transceiver array, which is exposed on the PeerConnection Unity
component.

Using customized CreateOffer() and SetRemoteDescriptionAsync() methods
overriding the behavior of the same-named methods found in the C#
library, the PeerConnection component is able to automatically pair the
transceivers and connect the associated audio and video sources, in a
way that allows the user to setup those components purely in the Unity
editor without having to write any specific code. The StandaloneDemo
scene demonstrate this feature by showing how 4 video transceivers can
be used to stream in parallel multiple video tracks, and the effect of
missing senders and/or receivers on this pairing.