Ignore TrackRemoved event with no MediaLine
Ignore TrackRemoved events associated with transceivers which do not
have a corresponding MediaLine in the PeerConnection component. These
are events for transceivers which are ignored due to the user not
configuring any Sender/Receiver component to handle them.
SetRemoteDescriptionAsync() already warns about this, so simply silently
ignore those events thereafter.