Rename M.MR.WebRTC.Native.dll into mrwebrtc.dll
Due to a bug in P/Invoke on Windows, which prevents loading DLLs by name
without extension when the filename contains some dot '.' character, and
in order to keep the Microsoft.MixedReality.WebRTC assembly fully
cross-platform, this change renames the native implementation DLL from
Microsoft.MixedReality.WebRTC.Native.dll into mrwebrtc.dll. This has the
added benefit to make it clear that this is not anymore the C++ library,
but an opaque internal implementation. The C++ library will, like the C#
one, build on top of that implementation DLL.