Update Unity tutorial for multi-track changes (#236)
Update the Unity tutorial in the documentation to reflect the latest API
changes introduced by the multi-track work (#152).

This change does not update the C++ and C# library documentations. This will be
done in a future change.

In addition, rename the CreateOffer() and SetRemoteDescriptionAsync() methods
of the PeerConnection Unity component to StartConnection() and
HandleConnectionMessageAsync(), respectively. This disambiguates them from the
methods of the PeerConnection class in the C# library, in an effort to
highlight that the former ones perform more work than the latter ones.