Handle transceiver kind mismatching in Unity
The Unity library uses a declarative model for transceiver pairing. This
means that due to user error the declaration on both peers might be
incompatible, in the sense that a same media line index can have an
audio media line on one peer and a video media line on the other peer.
This change detects the mismatching during the SDP exchange, and throws
a specific exception later logged to the Unity console, to make things
clear for the user and avoid hard-to-understand assertions down the line.