Upgrade .NET Core UWP to 6.2.10
Upgrade the Microsoft.NETCore.UniversalWindowsPlatform NuGet package to
version 6.2.10 for use in TestAppUWP to fix a bug in .NET Native AOT
toolchain failing to generate some code when marshaling structs passed
as `in` parameter.

Revert all workaround uses of `ref` to `in` where the input struct is
not expected to be read back, and is only passed one way.