Change projects to link library produced by WinRTC patches
Projects are changed to link the webrtc library produced by the patches at
https://github.com/microsoft/winrtc/tree/user/LoadLibrary/Arm64/patches_for_WebRTC_org/m80

Duplicate/obsolete sections in the vcxprojs have been tidied up.

Note that while the Win32 version of Microsoft.MixedReality.WebRTC.Native builds
using C++14, the UWP version uses C++17. This is because webrtc.lib is currently
build with C++17 for UWP but not for other configurations.