Add transceiver event raised when associated
Add a new transceiver event raised when the transceiver is associated
with a media line after an offer, local or remote, has been applied to
the peer connection.

The event is raised once the media line index has been updated to
reflect the index of the media line the transceiver is now associated
with.