Expose real media line index in transceiver
This change exposes the exact value of the media line index as used
internally in the transceiver when Unified Plan semantic is in use.
Previously the exposed index was incorrectly thought to be the index
into the transceiver array, which was correct in most cases because
transceivers are pushed into that array as they are created. However,
there are two things breaking that assumption:
- Data channels consume a media line in the SDP message, therefore the
  transceiver's media lines, which are only for audio and video, cannot
  possibly form a contiguous monotonic series of indices.
- If both peers add some transceivers before an offer is made by any of
  them, the callee receiving the offer will have duplicate "media line
  index" (as previously exposed), one set from the offer matching the
  one from the remote peer, and one set from the locally-created
  transceivers, which internally have no mid yet (not associated).

This change exposes the actual media line index (mid) of the transceiver
in an attempt to clarify handling by the user. The transceiver's media
line index is now invalid (-1) until the transceiver is associated to a
media line, as defined in the JSEP specification and referenced in the
WebRTC 1.0 standard.