Threadsafe PropertyChanged for CollectionViewModel
Make the CollectionViewModel of TestAppUWP thread-aware when raising the
PropertyChanged event. This avoids having to dispatch all internal
WebRTC callback handlings to the main UI thread, and only dispatch there
the event notification.

Note: This does not make the collection thread-safe, since its internal
container is not, although in practice it doesn't seem to cause issues.
This should anyway be addressed eventually.