Earlier and more consistent DataChannelAdded event
Raise the data channel added event earlier and more consistently for
out-of-band data channels when the data channel is created and added to
the peer connnection, instead of waiting for its state to change to
open, as the latter happens asynchronously much latter. This simplifies
interop wrapper creation, and makes the codepath consistent with in-band
data channels which already raise that event inline during the call to
AddDataChannel().