Retrieve actual mline index instead of mid
Change the implementation of ExtractMlineIndexFromRtpTransceiver() to
use the actual mline index retrieved from the internal implementation
instead of trying to deduce it from the mid value, because other
implementations than Google's may not use the mline index as the mid
value and this breaks interoperability. This is a workaround for the
fact the mline index is not exposed in the RtpTransceiverInterface.

Bug: #295