Split checkout and build PS scripts
Split the Powershell checkout and build scripts for Windows, and
properly handle the various build variants with a triple of arguments.

Note that this change disables ARM(32), which is not supported on this
branch due to MSVC ARM32 support being unavailable in Google's scripts
(see build/vs_toolchain.py:CopyDlls).