Move mrwebrtc-win32 and Android projects to C++14
The mrwebrtc library statically links with the Google libwebrtc
implementation, which normally doesn't support C++14. For UWP, the
WebRTC UWP SDK project adds C++17 support because this is needed for
C++/WinRT usage, but this is officially unsupported by Google. In order
to avoid binary mismatches on Win32 and Android, particularly between
std::optional and absl::optional, change those projects to use C++14
only. The mrwebrtc-uwp project stays on C++17 for now, with the custom
C++17 variant of libwebrtc for UWP.