Add fulldebug config to Android project (#341)
The default debug configuration on Android leaves debug symbols for
mrwebrtc, but strips them for private symbols, especially those in
libwebrtc. This makes debugging through that code difficult.

This change adds a fulldebug build variant which removes that stripping,
at the expense of producing a much larger Android archive (x5).