Revert mrwebrtc-win32 back to C++17 for now
The NuGet pacakges for the core prebuilt libwebrtc are still based on
C++17. Changing the project to C++14 would create a binary
incompatibility which breaks building for CI and all users using those
packages instead of recompiling libwebrtc from sources. So revert that
change until there are NuGet packages or similar that are compatible
with C++14. This does not change the Android project, which is
unaffected and needs C++14 for binary compatibility.