Init/close PeerConnection on enable/disable
Ensure proper shutdown across MediaLines & co.

Removed AutoInitializeOnStart, and made Initialize/Uninitialize private,
now PC should be turned on/off by enabling/disabling.

We now create/keep local tracks only when there are both a live source and a
connection. Therefore, local tracks do not persist after the PeerConnection
component is disabled. This is intended, as we cannot keep the track and
attach it to a different PeerConnection if the component is re-enabled.

Removed listeners to tracks removed. Unpairing is handled manually on
renegotiation and on shutdown (MediaLine.UnpairTransceiver).