Update the Unity library to handle split sources (#404)
Convert `WebcamSource` and `MicrophoneSource` to source objects shareable among
multiple peer connections.

Because the `SceneVideoSender` component uses the ARGB32 callback, also re-add
that callback, and add a TODO to remove all ARGB32 callbacks in favor of
explicit conversion utility functions, for clarity about the performance
implication of the conversion.

This change also splits the `VideoChatDemo` and `SceneCaptureDemo` for clarity,
leaving each demo with a single audio and video track per peer connection.  The
`SceneCaptureDemo` now has a placeholder showing the capture camera position.
Multiple tracks per peer connection are demo-ed in the `StandaloneDemo`.

All demos are also updated to show essential components as Unity objects in the
scene, instead of adding multiple of them on the same `GameObject`. This helps
with discoverability.