Disable device sources when exception occur
Catch exceptions when failing to create a device track source in
WebcamSource and MicrophoneSource components, and disable the component
before re-throwing the exception to get Unity to log it. This ensures
that the components are disabled for consistency.

This is a clean-up loosely related to #335, though the code triggering
that issue has been changed so the issue does not occur anymore.