Handle source enable/disable correctly
MediaLines now create a sender track when a source is attached *and* live,
and destroy it if any of these is false.

Note that we could delay track creation to when a connection is established,
but this is simpler to do, and local tracks should be lightweight so creating
them eagerly won't make much of a difference.