[Unity] Move utils into main library
Move the UniformColorVideoSource and LocalOnlySignaler into the main
Unity library. Those classes are used both by the main library tests and
by the samples package, and this is the simplest working way to make
things work, as classes in test assemblies cannot be used in runtime
ones (so can't use from samples package if in library's Tests/), and we
cannot add a dependency from the main library to the samples package.